Extract abstract policy base class.
